% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmenInteraction.R
\name{AugmenInteraction}
\alias{AugmenInteraction}
\title{Augmentation}
\usage{
AugmenInteraction(OrigriData, K)
}
\arguments{
\item{OrigriData}{An original binary attribute matrix with K columns}
}
\value{
The augmented matrix  'OrigriData_Aug' and the number of columns of 
OrigriData_Aug, 's_num'
}
\description{
Expanding the original binary latent attribute matrix to a matrix including 
all the interaction terms, which is required for "All_effect" and "DINA" cases
}
\examples{
A_test<-matrix(c(1,1,1,0,0,1,0,0),ncol=2,byrow=T)
A_test_Aug <- AugmenInteraction(A_test,2);
}
